# modules/common/swagger/api.yaml
swagger: "2.0"
info:
  description: "唐僧叨叨 API"
  version: "1.0.0"
  title: "唐僧叨叨 API"
host: "api.botgate.cn"
...
schemes:
  - "https"
basePath: "/v1"
paths:
  /manager/common/appconfig:
    get:
      tags:
        - "commonManager"
      summary: "获取app配置"
      description: "获取app配置"
      operationId: "manager common appconfig"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "返回"
          schema:
            properties:
              revoke_second:
                type: integer
                description: "消息撤回时间限制（单位秒）"
              welcome_message:
                type: string
                description: "登录欢迎语"
              new_user_join_system_group:
                type: integer
                description: "注册用户是否加入系统群 1.是"
              search_by_phone:
                type: integer
                description: "是否开启手机号搜索 1.是"
              register_invite_on:
                type: integer
                description: "是否开启注册邀请机制 1.开启"
              send_welcome_message_on:
                type: integer
                description: "是否开启注册登录发送欢迎语 1.开启"
              invite_system_account_join_group_on:
                type: integer
                description: "是否开启系统账号加入群聊 1.开启"
              register_user_must_complete_info_on:
                type: integer
                description: "是否开启注册用户必须填写完整信息 1.开启"
              channel_pinned_message_max_count:
                type: integer
                description: "频道置顶消息最大数量"
              can_modify_api_url:
                type: integer
                description: "是否允许修改api地址 1.允许"
              ip_white_list:
                type: string
                description: "ip白名单"
              login_type:
                type: string
                description: "登录类型"
              sensitive_words:
                type: string
                description: "敏感词"
              disable_change_device:
                type: integer
                description: "是否禁止更换设备 1.禁止"
              signup_device_limit:
                type: integer
                description: "设备限制注册限制数"
              sigle_ip_register_limit_in_12hour:
                type: integer
                description: "单IP12小时注册限制数"
              auto_clear_history_msg:
                type: integer
                description: "自动清除几天前历史消息"
              signup_invite_code_visible:
                type: integer
                description: "注册邀请码是否可见 1.是"
              signin_auth_code_visible:
                type: integer
                description: "登录授权码是否可见 1.是"
              friend_online_status_visible:
                type: integer
                description: "好友在线状态是否可见 1.是"
              mobile_msg_read_status_visible:
                type: integer
                description: "手机消息已读状态是否可见 1.是"
              mine_wallet_visible:
                type: integer
                description: "我的钱包是否开启 1.是"
              wallet_payout_min:
                type: integer
                description: "钱包提现最小金额"
              red_envelope_visible:
                type: integer
                description: "红包模块是否开启 1.是"
              transfer_visible:
                type: integer
                description: "转账模块是否开启 1.是"
              transfer_min_amount:
                type: integer
                description: "转账最小金额"
              mobile_edit_msg:
                type: integer
                description: "手机端是否可以编辑消息 1.是"
              group_member_see_member:
                type: integer
                description: "普通群成员是否可以查看其他群成员 1.是"
              sign_red_envelope_visible:
                type: integer
                description: "签到红包模块是否开启 1.是"
              msg_time_visible:
                type: integer
                description: "消息时间是否可见 1.是"
              pinned_conversation_sync:
                type: integer
                description: "置顶会话是否同步 1.是"
              only_internal_friend_add:
                type: integer
                description: "仅内部号可被加好友及加非内部号好友 1.是"
              only_internal_friend_create_group:
                type: integer
                description: "仅内部号可建群 1.是"
              only_internal_friend_send_group_red_envelope:
                type: integer
                description: "仅内部号可发群红包 1.是"
              only_internal_friend_send_group_card:
                type: integer
                description: "仅内部号可群内推送名片 1.是"
              only_internal_friend_group_robot_free_msg:
                type: integer
                description: "仅内部号群机器人免消息 1.是"
              group_member_limit:
                type: integer
                description: "群人数限制"
              user_agreement_content:
                type: string
                description: "用户协议内容"
              privacy_policy_content:
                type: string
                description: "隐私政策内容"
              moments_visible:
                type: integer
                description: "好友分享是否可见 1.是"  
        400:
          description: "错误"
          schema:
            $ref: "#/definitions/response"
      security:
        - token: []
    post:
      tags:
        - "commonManager"
      summary: "修改app配置"
      description: "修改app配置"
      operationId: "manager common appconfig update"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "req"
          description: "修改app配置信息"
          required: true
          schema:
            type: object
            properties:
              revoke_second:
                type: integer
                description: "消息撤回时间限制（单位秒）"
              welcome_message:
                type: string
                description: "登录欢迎语"
              new_user_join_system_group:
                type: integer
                description: "注册用户是否加入系统群 1.是"
              search_by_phone:
                type: integer
                description: "是否开启手机号搜索 1.是"
              register_invite_on:
                type: integer
                description: "是否开启注册邀请机制 1.开启"
              send_welcome_message_on:
                type: integer
                description: "是否开启注册登录发送欢迎语 1.开启"
              invite_system_account_join_group_on:
                type: integer
                description: "是否开启系统账号加入群聊 1.开启"
              register_user_must_complete_info_on:
                type: integer
                description: "是否开启注册用户必须填写完整信息 1.开启"
              channel_pinned_message_max_count:
                type: integer
                description: "频道置顶消息最大数量"
              can_modify_api_url:
                type: integer
                description: "是否允许修改api地址 1.允许"
              ip_white_list:
                type: string
                description: "ip白名单"
              login_type:
                type: string
                description: "登录类型"
              sensitive_words:
                type: string
                description: "敏感词"
              disable_change_device:
                type: integer
                description: "是否禁止更换设备 1.禁止"
              signup_device_limit:
                type: integer
                description: "设备限制注册限制数"
              sigle_ip_register_limit_in_12hour:
                type: integer
                description: "单IP12小时注册限制数"
              auto_clear_history_msg:
                type: integer
                description: "自动清除几天前历史消息"
              signin_auth_code_visible:
                type: integer
                description: "登录授权码是否可见 1.是"
              friend_online_status_visible:
                type: integer
                description: "好友在线状态是否可见 1.是"
              mobile_msg_read_status_visible:
                type: integer
                description: "手机消息已读状态是否可见 1.是"
              wallet_payout_min:
                type: integer
                description: "钱包提现最小金额"
              transfer_visible:
                type: integer
                description: "转账模块是否开启 1.是"
              transfer_min_amount:
                type: integer
                description: "转账最小金额"
              mobile_edit_msg:
                type: integer
                description: "手机端是否可以编辑消息 1.是"
              group_member_see_member:
                type: integer
                description: "普通群成员是否可以查看其他群成员 1.是"
              msg_time_visible:
                type: integer
                description: "消息时间是否可见 1.是"
              pinned_conversation_sync:
                type: integer
                description: "置顶会话是否同步 1.是"
              only_internal_friend_add:
                type: integer
                description: "仅内部号可被加好友及加非内部号好友 1.是"
              only_internal_friend_create_group:
                type: integer
                description: "仅内部号可建群 1.是"
              only_internal_friend_send_group_red_envelope:
                type: integer
                description: "仅内部号可发群红包 1.是"
              only_internal_friend_send_group_card:
                type: integer
                description: "仅内部号可群内推送名片 1.是"
              only_internal_friend_group_robot_free_msg:
                type: integer
                description: "仅内部号群机器人免消息 1.是"
              group_member_limit:
                type: integer
                description: "群人数限制"
              user_agreement_content:
                type: string
                description: "用户协议内容"
              privacy_policy_content:
                type: string
                description: "隐私政策内容"
      responses:
        200:
          description: "返回"
          schema:
            $ref: "#/definitions/response"
        400:
          description: "错误"
          schema:
            $ref: "#/definitions/response"
      security:
        - token: []
  /manager/common/appmodule:
    get:
      tags:
        - "commonManager"
      summary: "获取app模块"
      description: "获取app模块"
      operationId: "manager common get appmodule"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "返回"
          schema:
            type: array
            items:
              properties:
                sid:
                  type: string
                  description: "模块唯一ID"
                name:
                  type: string
                  description: "模块名称"
                desc:
                  type: string
                  description: "模块说明"
                status:
                  type: integer
                  description: "模块状态 1.可选 0.不可选 2.选中不可编辑"
        400:
          description: "错误"
          schema:
            $ref: "#/definitions/response"
      security:
        - token: []
    put:
      tags:
        - "commonManager"
      summary: "修改app模块"
      description: "修改app模块"
      operationId: "manager common appmoudle update"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "req"
          description: "修改app模块信息"
          required: true
          schema:
            type: object
            properties:
              sid:
                type: string
                description: "模块ID"
              name:
                type: string
                description: "模块名称"
              desc:
                type: string
                description: "模块说明"
              status:
                type: integer
                description: "模块状态 1.可选 0.不可选 2.选中不可编辑"
      responses:
        200:
          description: "返回"
          schema:
            $ref: "#/definitions/response"
        400:
          description: "错误"
          schema:
            $ref: "#/definitions/response"
      security:
        - token: []
    post:
      tags:
        - "commonManager"
      summary: "添加app模块"
      description: "添加app模块"
      operationId: "manager common appmoudle add"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "req"
          description: "新增app模块信息"
          required: true
          schema:
            type: object
            properties:
              sid:
                type: string
                description: "模块ID"
              name:
                type: string
                description: "模块名称"
              desc:
                type: string
                description: "模块说明"
              status:
                type: integer
                description: "模块状态 1.可选 0.不可选 2.选中不可编辑"
      responses:
        200:
          description: "返回"
          schema:
            $ref: "#/definitions/response"
        400:
          description: "错误"
          schema:
            $ref: "#/definitions/response"
      security:
        - token: []
  /manager/common/{sid}/appmodule:
    delete:
      tags:
        - "commonManager"
      summary: "删除app模块"
      description: "删除app模块"
      operationId: "manager common appmoudle delete"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "sid"
          type: string
          description: "模块ID"
          required: true
      responses:
        200:
          description: "返回"
          schema:
            $ref: "#/definitions/response"
        400:
          description: "错误"
          schema:
            $ref: "#/definitions/response"
      security:
        - token: []
...
  /common/appversion/{os}/{version}:
    get:
      tags:
        - "common"
      summary: "获取最新版本"
      description: "获取最新版本"
      operationId: "check new version"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "os"
          type: string
          description: "平台类型android|ios"
          required: true
        - in: "path"
          name: "version"
          type: string
          description: "版本号"
          required: true
      responses:
        200:
          description: "返回"
          schema:
            properties:
              app_version:
                type: string
                description: "版本号"
              os:
                type: string
                description: "平台 android|ios"
              is_force:
                type: integer
                description: "是否强制更新 1.是"
              update_desc:
                type: string
                description: "更新说明"
              download_url:
                type: string
                description: "下载地址"
              created_at:
                type: string
                description: "更新时间"
        400:
          description: "错误"
          schema:
            $ref: "#/definitions/response"
      security:
        - token: []
...
  /common/appmodule:
    get:
      tags:
        - "common"
      summary: "app模块列表"
      description: "app模块列表"
      operationId: "appmodule"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "返回"
          schema:
            type: array
            items:
              properties:
                sid:
                  type: string
                  description: "模块ID"
                name:
                  type: string
                  description: "模块名称"
                desc:
                  type: string
                  description: "模块说明"
                status:
                  type: integer
                  description: "模块状态 1.可选 0.不可选 2.默认选中不可编辑"

        400:
          description: "错误"
          schema:
            $ref: "#/definitions/response"
      security:
        - token: []
...
securityDefinitions:
  token:
    type: "apiKey"
    in: "header"
    name: "token"
    description: "用户token"
definitions:
  response:
    type: "object"
    properties:
      status:
        type: integer
        format: int
      msg:
        type: "string"